README
This automated analysis package is a summary of my learning experience of  single-cell transcriptome. It contains a large number of built-in functions for analysis, allowing for streamlined completion of the entire single-cell transcriptome analysis in just 1-2 lines of commands. Additionally, it includes visually appealing visualization plots and can generate reports summarizing key findings. This current version, version 1, is implemented entirely in the R programming language. I will continue to update and add more analysis applications in future versions. Your feedback and suggestions are welcome and appreciated.